api/requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
ffmpeg-python
psutil
google-generativeai
pydantic
